+++ changelog.md
@@ -1,3 +1,5 @@
 ## 1.0.0 - 2019-06-20 - CodeIniter - 3.1.10
 * initial plugin
 * add module Pekerjaan
+* add module Agama
+* add module Pendidikan